Safari-609.1.11.1/LayoutTests/js/dom/unhandled-promise-rejection-basic-in-workers-expected.txt

[Worker] Test unhandled promise rejection event.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Starting worker: ../resources/unhandled-promise-rejection-basic-in-workers.js
PASS [Worker] error.type is "unhandledrejection"
PASS [Worker] error.cancelable is true
PASS [Worker] error.promise is promise
PASS [Worker] error.reason is "ERROR"
PASS successfullyParsed is true

TEST COMPLETE
